+++ client.h
@@ -23,8 +23,8 @@ public:
     int64_t sendMessage(Message* m);
     int32_t getUsablePacketID();
     int ackMessage(uint16_t pID);
-    int connect(const std::string addr, int port, bool cleanSession);
-    int publish(const std::string topic, const std::string data, uint8_t qos, bool retain);
+    int64_t connect(const std::string addr, int port, bool cleanSession);
+    int64_t publish(const std::string topic, const std::string data, uint8_t qos, bool retain);
 };
